+++ src/ui/MessageFeedbackModal/index.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables';
+@import '../../styles/variables';
 
 .sendbird-message-feedback-modal-content__mobile {
   @include mobile() {
